Resolve "Upgrade to Tendermint v0.33.4"

Closes #146

See merge request thorchain/midgard!253

+++ internal/models/refund.go
@@ -12,7 +12,7 @@ type EventRefund struct {
 
 func NewRefundEvent(refund types.EventRefund, event types.Event) EventRefund {
 	return EventRefund{
-		Code:   uint32(refund.Code),
+		Code:   refund.Code,
 		Reason: refund.Reason,
 		Event:  newEvent(event),
 	}
